requirements.txt
gradio>=4.0.0
fpdf==1.7.2
pdfplumber==0.10.2
python-docx==1.1.0
langchain>=0.1.0
langchain-community>=0.0.30
langchain-ollama>=0.1.0
ollama>=0.1.0
